Bug: 22290 Ajuste desfazendo a conversão de tipo de código de aluno ao obter dados de frequência.

+++ src/SME.SGP.Dados/Repositorios/RepositorioFrequenciaAlunoDisciplinaPeriodo.cs
@@ -22,7 +22,7 @@ namespace SME.SGP.Dados.Repositorios
                         from
 	                        frequencia_aluno
                         where
-	                        codigo_aluno = @codigoAluno::int8
+	                        codigo_aluno = @codigoAluno
 	                        and disciplina_id = @disciplinaId
 	                        and tipo = @tipoFrequencia
 	                        and periodo_escolar_id = @periodoEscolarId";
@@ -85,7 +85,7 @@ namespace SME.SGP.Dados.Repositorios
             var query = new StringBuilder(@"select * 
                             from frequencia_aluno
                            where tipo = 2
-	                        and codigo_aluno = @alunoCodigo::int8
+	                        and codigo_aluno = @alunoCodigo
                             and turma_id = @turmaCodigo ");
 
             if (!string.IsNullOrEmpty(componenteCurricularCodigo))
@@ -103,7 +103,7 @@ namespace SME.SGP.Dados.Repositorios
         public async Task<IEnumerable<FrequenciaAluno>> ObterFrequenciaBimestresAsync(string codigoAluno, int bimestre, string codigoTurma)
         {
             var query = @"select * from frequencia_aluno fa 
-                            where fa.codigo_aluno = @codigoAluno::int8
+                            where fa.codigo_aluno = @codigoAluno
                             and fa.turma_id = @turmaId and fa.tipo = 1";
 
             if (bimestre > 0)
@@ -123,7 +123,7 @@ namespace SME.SGP.Dados.Repositorios
         {
             var query = new StringBuilder(@"select *
                         from frequencia_aluno
-                        where codigo_aluno = @codigoAluno::int8
+                        where codigo_aluno = @codigoAluno
 	                        and tipo = @tipoFrequencia
 	                        and bimestre = @bimestre ");
 
@@ -146,7 +146,7 @@ namespace SME.SGP.Dados.Repositorios
                         from frequencia_aluno fa
                         inner join periodo_escolar pe on fa.periodo_escolar_id = pe.id
                         where
-	                        codigo_aluno = @codigoAluno::int8
+	                        codigo_aluno = @codigoAluno
 	                        and tipo = @tipoFrequencia
 	                        and pe.periodo_inicio <= @dataAtual
 	                        and pe.periodo_fim >= @dataAtual ");
@@ -168,7 +168,7 @@ namespace SME.SGP.Dados.Repositorios
             var query = @"select *
                         from frequencia_aluno fa
                         inner join periodo_escolar pe on fa.periodo_escolar_id = pe.id
-                        where codigo_aluno = @codigoAluno::int8
+                        where codigo_aluno = @codigoAluno
                             and disciplina_id = @disciplinaId
 	                        and tipo = 1
 	                        and pe.periodo_inicio <= @dataAtual
